Workaround for compiler warning (-Wformat) 32bit vs. 64bit gcc
 - Complement to commit 2949663d2b16f2b6e7b7a195275cc8bc53caf1ee

+++ src/mv2mysql.cpp
@@ -192,7 +192,7 @@ fwrite(jsonBuf.c_str(), jsonBuf.length(), 1, f);
 fclose(f);
 /* ################################################ */
 
-	printf("done (%lu Bytes)\n", fsize); fflush(stdout);
+	printf("done (%u Bytes)\n", (uint32_t)fsize); fflush(stdout);
 
 	return true;
 }
@@ -290,7 +290,7 @@ bool CMV2Mysql::parseDB(bool /*is_fulldb*/)
 	}
 	videoInfo.push_back(videoInfoEntry);
 
-	printf("done (%lu entrys)\n", videoList.size()); fflush(stdout);
+	printf("done (%u entrys)\n", (uint32_t)(videoList.size())); fflush(stdout);
 	jsonBuf.clear();
 	return true;
 }
